+++ pkg/daemon/sysproc_unix.go
@@ -1,0 +1,14 @@
+//go:build !windows
+
+package daemon
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func setDetached(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{
+		Setsid: true,
+	}
+}