+++ pkg/daemon/sysproc_windows.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package daemon
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func setDetached(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{
+		CreationFlags: 0x08000000, // DETACHED_PROCESS
+	}
+}